+++ internal/server/server.go
@@ -42,7 +42,7 @@ import (
 )
 
 var (
-	Version     = "1.5.1"
+	Version     = "1.6.0"
 	connCounter uint64
 )